rename and move a bunch of things
